refactor dto namespace
