+++ DTOs/UserDTOs/UserDTO.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace DTOs.UserDTOs.UserDTO
+namespace SkillStackCSharp.DTOs.UserDTOs
 {
     public class UserDTO
     {
